Do not create python bytecode files for tests
Python usually compiles the source code files into byte code before or
during execution and stores them on disk so as to not perform redundant
compilations multiple times. For Python 3, the __pycache__ directories
are created in all packages which contain the pre-compiled files. In
Python 2, each compiled source file is directly accompanied by a
*.pyc/*.pyo file.
In general, one goal of all my projects is to be as stateless (in all
senses) as possible. Leaving no newly created files after running the
test suite is one aspect of this aim. To that end, this change disables
storing of these byte-code files by setting the PYTHONDONTWRITEBYTECODE
environment variable before invoking the interpreter.

Notes: According to my information this approach covers Python 2
       (starting with version 2.6) and Python 3 alike [1].

[1] https://docs.python.org/2/using/cmdline.html#envvar-PYTHONDONTWRITEBYTECODE