Add support for listing available snapshots
As it stands, it is not possible to query all available snapshots or
list ones that were just created. To create incremental snapshots,
however, it is of paramount importance to know which existing ones can
be used as a basis.
This change introduces the functionality to list all snapshots present
below a given directory.
With this change we also introduce the notion of a repository. A
repository is a directory that contains or is supposed to contain
snapshots of an arbitrary number of subvolumes. In the future a
repository will form the basis for various backup related operations.