Allow for correct snapshot listing relative to repository root
The previous change determined that snapshot listing does not work
properly when performed on a sub-directory. Consider the following
directory tree:
  .
  ├── subvolume
  └── snapshots
      └── snapshot

When invoking 'btrfs subvolume -o snapshots/' (a command very similar to
the one that is actually used to generate the list of snapshots in a
programmatical fashion), it will return the path 'snapshots/snapshot'
and *not* simply 'snapshot' -- regardless of the actual path given to
the command ('snapshots/' in the example) or the current working
directory.
As it turns out, working around this problem is all but trivial. The
found solution relies on detection of the btrfs root (as an absolute
path) and concatenating the returned snapshot path with the btrfs root
path. Afterwards, the (absolute) path to the actual directory for which
the call was made can just be removed from the now absolute snapshot
path -- it is a common prefix.
This change puts in place the infrastructure to find the btrfs root
given an arbitrary directory and adjusts the Repository class to use it
to generate snapshot paths that are relative to the repository's root
directory.