Introduce subvolume diff support
The previously introduced support for querying existing snapshots is
only half the battle when it comes to incremental snapshots. Another
factor that comes into play is whether or not to create a new snapshot
in the first place.
This change introduces support for listing the changed files in a btrfs
subvolume with respect to a present snapshot. That is, if a file in the
subvolume got changed after a snapshot was taken it is reported as such.
On the other hand, this mechanism also helps in determining whether or
not a new snapshot is required in the first place.
Note that although snapshotting on btrfs is cheap because of its
copy-on-write nature, a snapshot that is completely superfluous is still
considered pointless and should be avoided. This change helps in
achieving this goal.