Verify that different locales sort our time stamp values as intended
We use a time stamp format of the form YYYY-MM-DD_hh:mm:ss for naming our
snapshots. This is also the format that the btrfs command uses in
various places (when listing snapshots including their creation times,
for instance).
For correct functioning of the program it is important that sorting time
stamps of this format always yields the same order in the sense that
older time stamps emerge earlier and more recent ones are included at
the end.
The behavior of string sortings is potentially influenced by the locale
currently in use. Even if Python does not include the locale into
standard string comparison in a list (which I am not 100% sure), the
btrfs tool might. This change adds a test case that checks for
compliance with all locales available on the system.