Do not pass -s to subvolume list btrfs command invocation
As it turns out, the property of a snapshot being a snapshot is lost
once we send it somewhere. That is, such a snapshot will not appear in
the list of available snapshots -- it is considered an ordinary (but
read-only subvolume).
There is one way to help preserve the snapshot property, however. That
is by means of the clone-src option (-c). This option accepts the name
of a snapshot which it expects to be available on the source and the
destination. The send operation can then reduce the amount of data
transferred.
We cannot use the -c option because we cannot guarantee that a snapshot
is present on the destination. For example, when we transfer the very
first snapshot of a subvolume the destination is empty, so there is no
chance to make use of this option.
To that end, a different approach is taken to ensure a snapshot listing
will contain all snapshots: we do not require the snapshot property.
That is, we remove the -s flag from the command invocation. The downside
of this approach is that we not only list snapshots but also read-only
subvolumes. Since we have dedicated repositories which are generally
considered to be under our control (and we do not create read-only
subvolumes), this approach is considered safe.