Append a number to a snapshot's name if it is not unique
If repositories are synchronized in very rapid succession (while data is
being added), it might happen that a snapshot is created in the same
second a previous one was. A name clash would be the result and the
btrfs snapshot command would not succeed.
This change introduces logic to cover these (unlikely) cases. After a
snapshot name is created, we check whether it is unique. If that is not
the case, we append a number (preceeded by a dash) and try again. With
each failed attempt, the number is incremented.