Add support for incremental synchronization
This change finally adds support for incremental synchronization. The
additions are fairly straight forward. We already retrieved the path to
the most recent snapshot. This is the perfect candidate to use as a
parent. So all we have to do is to check whether it exists on the
destination repository. If that is not the case, we have to perform a
full sync, otherwise an incremental transfer of the changes since the
most recent snapshot is sufficient.