Unify test harness creation functionality
In order to test btrfs snapshot or repository related functionality, a
test harness of subvolumes, files, and directories is used most of the
time. Test creation is somewhat tedious, however, and test cases are not
as concise as desired.
This change marks a first step towards changing the status quo a bit. It
introduces a 'make' function that can be used to create files,
directories, and even subvolumes with a single call. The function
creates directories as necessary before creating the desired target
object, i.e., it is no longer necessary to explicitly create all
directories below a given file.