Add --reverse option
When restoring data from a backup the source and the destination
repositories are logically swapped over the creation case. That is, if
you backed up your data from /snapshots to /backup, for instance, you
would restore it from /backup to /snapshots. This behavior is logical
and intuitive.
However, sometimes it is more easily to have the program deal with this
swapping of the parameters in order to not have to create an entirely
new command line with these two parameters reversed. For that matter,
this change introduces the --reverse option that essentially swaps the
source and the destination repository.
Note that although not intended, the newly introduced option can also be
used in the non-restore case, i.e., when creating a new snapshot in the
source directory and synchronizing it with a destination repository.