Amend command line arguments
Ever since the introduction of the restore functionality we had problems
with the inequality between the backup and the restore paths: backing up
is the default and restoring has to be enabled explicitly. Another point
of critique is the increasing number of options allowed for both
commands, most of them being mutually exclusive (which was not
enforced).
This change adjusts the approach taken for command line option parsing.
Instead of relying on a specific option for restoration (over the
default of backing up) we make them both first class citizens by
introduction of respective sub-commands. Using this approach provides a
simple way of restricting mutually exclusive options between the two
workings while allowing for more expressive and yet customized help
texts. In addition to this advantage, we do not have to fiddle around
with ArgumentParsers' add_mutually_exclusive_group method which appears
to be very restrictive in how it can be combined with other command
groups, for example.