Fix lowercase 'usage:' string in help text
The default HelpFormatter class used for formatting the usage string
presented to the user when he/she invokes the program with the -h/--help
option is written in lowercase. This casing is not in line with the
other strings presented in the help text.
This change fixes the ugliness by providing a custom HelpFormatter
subclass to lay out the usage. One problem encountered was that
subparsers duplicate the usage prefix for some reason when not used
carefully. To that end, two classes had to be provided, one for the
top-level parser and another for the sub-parsers.