Introduce remote command execution support
Arguably, in a fair share of cases one would want to backup data not to
a local device but rather a remote entity. One possibility for remotely
accessing file systems is by mounting them locally using NFS, for
example. However, once mounted over NFS, certain file system properties
are abstracted from. The particularities of a btrfs file system, for
instance, are not exposed through NFS (and similar approaches). This
constraint renders any such mounting related approach effectively
unusable for btrfs-backup.
In order to still be able to backup data remotely, this change
introduces support for running commands remotely instead of only
locally. Since btrfs-backup is basically a sophisticated command
execution engine around the btrfs tool suite, we can execute the
commands not only locally but also remotely, using SSH, for instance
(this fact of course implies that the remote side has a btrfs file
system available along with the btrfs programs).
With this change it is now possible to specify a remote command to use
by means of the newly introduced --remote-cmd option. The command will
be used to prefix all actually performed btrfs commands which means that
the remote command must accept the command to execute remotely as its
argument. Note that although the approach was designed with SSH in
mind, really any command that allows for connecting to a remote host and
running a command should work.