Introduce --debug option
Sometimes it is useful to get a fully blown backtrace instead of just a
single error message along with an error code that tries to describe the
problem that occurred.
This change introduces a new option, --debug, that essentially causes
btrfs-backup to not catch any exceptions on the top level. Since
Python's default behavior is to print the full exception backtrace if an
exception escapes the program, we will get the desired additional
information.
This explanation is not the full story, however, otherwise we could just
print the backtrace ourselves. The other use case is testing, where we
sometimes also require an exception to escape the main function for
various reasons. Both cases are covered by introduction of this new
option.