Add support for send and receive filtering
An additional future goal of btrfs-backup is to support end-to-end
encryption of data to backup. That is, it must support encryption of
outgoing data, storing this data on a remote site, and restoring it from
there while decrypting it.
This goal will be achieved by pipelining the output of the btrfs send
command through an encryption program as a first step. Later, in the
restore phase, the stored encrypted stream will be read back and
decrypted once it is received. Afterwards, the snapshot/subvolume can be
remanifested accordingly.
An intermediate step to achieve this goal is to introduce general
filtering capabilities, i.e., the insertion of arbitrary pipeline
compatible commands into the send & receive pipeline. This change
introduces the --send-filter and --recv-filter options that accept such
filtering commands and applies them to the stream of data that is backed
up or restored, respectively.
In itself this functionality is only of limited use right now. Although
scenarios involving compression are possible with these new options, it
largely depends on the setup and the systems involved whether this
proves advantageous over a filter less approach.