Adhere to new root find behavior introduced with btrfs-progs 3.17.3
Starting with btrfs-progs 3.17.3 the subvolume show command not only
prints an error message but also returns an error code in case the
supplied directory is no btrfs subvolume. This change in behavior causes
btrfs-backup to function incorrectly.
With this change the _isRoot function is adjusted to accomodate for such
failures and to indicate that the given directory is no btrfs root. It is
worth pointing out that until now we used the failure of btrfs show to
guard against creation of Repository objects in non-existant
directories. Because of the mentioned change in behavior this error is
now masked and we can no longer rely on it being propagated properly. To
that end, this change furthermore adds an explicit check for the
existence of the given directory to _isRoot and if the directory does
not exist an error is raised.