Prepare for repository hierarchy introduction
A subsequent change will introduce a small repository class hierarchy.
In preparation for that change we refactor the code a bit. Specifically,
this change splits the snapshot method into two distinct steps after
retrieval of the list of btrfs root relative snapshots: the conversion
into a list of absolute snapshots and the conversion into a list of
repository relative snapshots. The latter functionality will be used in
other parts as well.
Furthermore, the synchronization of the brtfs volume contents with the
disk is moved from the deployment to the creation of a snapshot because
at the creation we can be 100% sure to have a btrfs subvolume which
actually supports this operation (for deployment this guarantee cannot
be given with upcoming changes).