Factor out RepositoryBase class
An upcoming change will introcuce a different type of repository that
can handle non-native snapshots, that is, snapshots that reside on a
non-btrfs file system in the form of a file (as opposed to a subvolume
as currently supported).
To prepare for this step, this change refactors the code and introduces
a repository base class which will serve as the foundation for the two
repositories by providing a common interface.