Change parameter passing method to program objects
With the upcoming introduction of another repository type we need to
create repositories not as part of the initialization of a Program
object but rather in the method being invoked afterwards to start the
backup or restore.
This change refactors the parameter passing from the main module to the
program object slightly to prepare for this introduction: we directly
treat the namespace object used during parameter parsing as a dict and
pass in all its values as keyword arguments. In order to allow for this
process to work as expected, unrequired (and already processed) fields
are removed from the namespace beforehand.