Introduce file repository support
In order to support enhanced filtering techniques on snapshot streams,
we can no longer remanifest a btrfs snapshot or subvolume from its
stream under all circumstances. In cases when encryption is involved,
btrfs cannot handle the data in any meaningful way, for example.
Instead, we have to persist the (potentially processed) stream into a
file and in order to restore the subvolume later on, we have to read
from the file, apply the filters in reverse order, and remanifest the
subvolume.
This change introduces another type of repository that provides the
ability to handle snapshots persisted in files: the FileRepository. File
repositories are limited in the functionality they offer. For instance,
it is not possible to directly create a snapshot on it. As such, they
are meant to be used only as source or destination repository for file
transfers. More specific, in case of a backup action, a file repository
may be the destination. If a restore operation is in progress, a file
repository may be used as the source. File repositories are always used
on conjunction with "real" repositories because of the way snapshotting
and backup on btrfs works: a snapshot is created on the local repository
and, if desired, it is transferred somewhere (potentially in an
incremental way with only the changes from a previous snapshot being
sent).
Another advantage of file repositories is their independence of the
underlying file system. Native repositories require a btrfs file system
to operate on. When storing data in an ordinary file, any file system
can be used.